Logging, enums, cosmetics
git-svn-id: svn+ssh://svn.yandex.ru/yt/trunk@905 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ yt/ytlib/bus/packet.cpp
@@ -9,9 +9,10 @@ static NLog::TLogger& Logger = BusLogger;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-const int THeaderTraits<TPacketHeader>::FixedSize = sizeof (TPacketHeader);
+const int THeaderTraits<TPacketHeader>::FixedSize =
+    sizeof (TPacketHeader);
 const int THeaderTraits<TMultipartPacketHeader>::FixedSize =
-        (size_t) &(((TMultipartPacketHeader*) NULL)->PartSizes);
+    (int) &(((TMultipartPacketHeader*) NULL)->PartSizes);
 
 ////////////////////////////////////////////////////////////////////////////////
